+++ app/src/main/java/in/teacher/dao/TempDao.java
@@ -30,7 +30,7 @@ public class TempDao {
             t.setExamId(c.getLong(c.getColumnIndex("ExamId")));
             t.setActivityId(c.getLong(c.getColumnIndex("ActivityId")));
             t.setSubActivityId(c.getLong(c.getColumnIndex("SubActivityId")));
-            t.setStudentId(c.getInt(c.getColumnIndex("StudentId")));
+            t.setStudentId(c.getLong(c.getColumnIndex("StudentId")));
             t.setSubjectId(c.getInt(c.getColumnIndex("SubjectId")));
             t.setSlipTestId(c.getLong(c.getColumnIndex("SlipTestId")));
             t.setSyncTime(c.getString(c.getColumnIndex("SyncTime")));
@@ -107,7 +107,7 @@ public class TempDao {
         sqliteDatabase.update("temp", cv, "id=1", null);
     }
 
-    public static void updateStudentId(int studentId, SQLiteDatabase sqliteDatabase) {
+    public static void updateStudentId(long studentId, SQLiteDatabase sqliteDatabase) {
         ContentValues cv = new ContentValues();
         cv.put("StudentId", studentId);
         sqliteDatabase.update("temp", cv, "id=1", null);
